pyproject.toml
[tool.poetry]
name = "pixelpalette"
version = "0.1.0"
description = "A web application for generating AI art using DALL-E"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 